// src/app/models/Patient.ts
export interface Patient {
    identificationType:   string;
    identificationNumber: string;
    firstName:            string;
    lastName:             string;
    dateOfBirth:          string;
    phone:                string;
    email:                string;
    password:             string;
    confirmEmail:         string;
    confirmPassword:      string;
    userId: string;
}